pyproject.toml
# ===================================================================
# pyproject.toml v2.0 (Canonical Poetry Standard)
#
# Unified project file, converted to native Poetry format
# for maximum clarity and robustness.
# ===================================================================

[tool.poetry]
name = "Quantization-Aware-Pre-Training"
version = "0.1.0"
description = "Experiments for AccML 2026 on quantized aware pre-training."
authors = ["Xavier Pillet <xavier.pillet@ls2n.fr>"]
license = "MIT"
readme = "README.md"
packages = [{include = "Quantization-Aware-Pre-Training", from = "src"}] # Spécifie une structure de type 'src'

[tool.poetry.dependencies]
python = "3.10.4"
accelerate = "==0.30.1"
datasets = "^2.19.0" 
pyarrow = ">=15.0"
sentencepiece = "0.1.97"
protobuf = "3.20.1"
evaluate = "0.2.2"
tensorboard = "2.11.0"
torch = ">=1.3"
transformers = { version = "4.38.0", extras = ["torch"] }
numpy = "<2.0"
pyyaml = "^6.0" 
gitignore-parser = "^0.1.13"
termcolor = "^3.1.0"
python-dotenv = "^1.1.1"
scikit-learn = "^1.7.2"
pandas = "^2.3.3"
codecarbon = "^2.1.4"
matplotlib = "^3.10.7"
gdown = "^5.2.0"
nltk = "^3.9.2"
sentence-transformers = "^3.0.0"
seqeval = "^1.2.2"
pytorch-lightning = "^2.0.0"
librosa = "^0.4.1"
soundfile = "^0.13.1"
seaborn = "^0.13.2"
lxml = "^6.0.2"
nvidia-ml-py = "^13.580.82"

[tool.poetry.group.dev.dependencies]
pyyaml = "^6.0"
pytest = "^8.0" # Version stable actuelle

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"